backend/requirements.txt
fastapi
uvicorn
google-generativeai
crewai
crewai-tools
python-dotenv
psycopg2-binary
pydantic-settings